ct: Don't even try to kill container w/o pidns
This is racy and hard to do right (if at all possible). Thus
just ban this for now.

+++ src/ct.c
@@ -268,6 +268,8 @@ int libct_container_kill(ct_handler_t h)
 
 	if (ct->state != CT_RUNNING)
 		return -1;
+	if (!(ct->nsmask & CLONE_NEWPID))
+		return -1;
 
 	kill(ct->root_pid, SIGKILL);
 	return 0;
